sync

+++ app/src/main/java/com/slash/youth/domain/FreeTimeServiceBean.java
@@ -41,7 +41,7 @@ public class FreeTimeServiceBean {
         public static class ListBean {
             private String avatar;
             private long endtime;
-            private int id;
+            private long id;
             private int instalment;
             private int isauth;
             private double lat;
@@ -54,7 +54,7 @@ public class FreeTimeServiceBean {
             private long starttime;
             private int timetype;
             private String title;
-            private int uid;
+            private long uid;
 
             public String getAvatar() {
                 return avatar;
@@ -72,11 +72,11 @@ public class FreeTimeServiceBean {
                 this.endtime = endtime;
             }
 
-            public int getId() {
+            public long getId() {
                 return id;
             }
 
-            public void setId(int id) {
+            public void setId(long id) {
                 this.id = id;
             }
 
@@ -176,11 +176,11 @@ public class FreeTimeServiceBean {
                 this.title = title;
             }
 
-            public int getUid() {
+            public long getUid() {
                 return uid;
             }
 
-            public void setUid(int uid) {
+            public void setUid(long uid) {
                 this.uid = uid;
             }
         }
